pyproject.toml
[project]
name = "pipeline-spotify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.7",
]

[tool.ruff]
line-length = 88
indent-width = 4

lint.select = [
    "E",
    "F",
    "W",
    "C90",
    "I",
    "UP",
    "B",
    "DTZ",
    "C4",
    "RET",
    "ARG",
    "N",
    "S",
]

lint.ignore = [
    "S101",
]

# Max complexity
lint.mccabe.max-complexity = 10

# Docstrings
lint.pydocstyle.convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
